requirements.txt
django>=3.0,<4.0
django-rest-auth>=0.9.5
djangorestframework>=3.12.2
djangorestframework-jwt>=1.11.0